first pass at integration with fyne gui
